feat: centralize server supabase client
